Change output layer to use sigmoid function
In theory, using a softmax on two classes is identical to using a
sigmoid on a flat array. The advantage of having a sigmoid is that there
are half as many parameters, therefore the train time might be decreased.

To this end, the structure of the neural nets in cnn.py and snn.py have
been altered to have this new structure. In addition, the reading in of
data in waveform_dataset.py has been altered to return flattened, not
one-hot, labels, and the custom metrics have been altered to accommodate
this.

Before merging, it would be wise to split the metrics and data-loading
code off into its own separate folder, or find some other way of
accommodating both sets of code in the same methods, as the methods in
the current master branch are required to reload the models from the
model store.