Modify class proportion decay to make it work
This commit modifies the class proportion decay to make it work
correctly and implements it in the neural networks. NB - this commit
must be run with the test flag until modifications are made to the model
store summary csv file. This commit also outputs debugging information
to ensure that rate decay does/does not take place when desired.